Super Admin Now working as intended
