+++ ASI.Basecode.WebApp/Authentication/SignInManager.cs
@@ -88,6 +88,7 @@ namespace ASI.Basecode.WebApp.Authentication
 
                 new Claim("UserId", user.UserId, ClaimValueTypes.String, Const.Issuer),
                 new Claim("UserName", user.Name, ClaimValueTypes.String, Const.Issuer),
+                new Claim("Role", user.Role, ClaimValueTypes.String, Const.Issuer),
             };
             return new ClaimsIdentity(claims, Const.AuthenticationScheme);
         }
