+++ ASI.Basecode.WebApp/Controllers/SuperAdminController.cs
@@ -8,14 +8,110 @@ using ASI.Basecode.Data.Models;
 using ASI.Basecode.Services.ServiceModels;
 using ASI.Basecode.Services.Interfaces;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 
 namespace ASI.Basecode.WebApp.Controllers
 {
     public class SuperAdminController : Controller
     {
-        public IActionResult SuperAdminDashboard()
+        private readonly IUserService _userService;
+        public SuperAdminController(IUserService userService)
+        {
+            _userService = userService;
+        }
+
+        // GET: UsersController
+        public ActionResult UserList()
+        {
+            var users = _userService.GetAllUsers(); // Fetch users from the service
+            return View(users);
+        }
+
+        // GET: UsersController/UserAdd
+        public ActionResult UserAdd()
         {
             return View();
         }
+
+        // GET: UsersController/Edit/UserId
+        public ActionResult Edit(string id)
+        {
+            var user = _userService.GetUserById(id);
+            if (user == null)
+            {
+                return NotFound();
+            }
+            return View(user);
+        }
+
+        // POST: SuperAdminController/Edit/userId
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult Edit(User user)
+        {
+            try
+            {
+                if (ModelState.IsValid)
+                {
+                    _userService.UpdateUser(user);
+                    return RedirectToAction(nameof(UserList));
+                }
+                return View(user);
+            }
+            catch
+            {
+                return View(user);
+            }
+        }
+
+        // POST: SuperAdminController/ResetPassword/userId
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult ResetPassword(string id)
+        {
+            try
+            {
+                _userService.ResetPassword(id, "Temp_123");
+                return RedirectToAction(nameof(UserList));
+            }
+            catch
+            {
+                return View();
+            }
+        }
+
+        // GET: SuperAdminController/Delete/userId
+        public ActionResult Delete(string id)
+        {
+            try
+            {
+                return RedirectToAction(nameof(UserList));
+            }
+            catch
+            {
+                return View();
+            }
+        }
+
+        // POST: UsersController/Delete/userId
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult Delete(int id, IFormCollection collection)
+        {
+            try
+            {
+                _userService.DeleteUser(id.ToString());
+                return RedirectToAction(nameof(UserList));
+            }
+            catch
+            {
+                return View();
+            }
+        }
+    public IActionResult SuperAdminDashboard()
+        {
+            return View();
+        }
+
     }
 }
