--- ASI.Basecode.WebApp/Controllers/UserController.cs
@@ -1,109 +1,0 @@
-﻿using ASI.Basecode.Services.ServiceModels;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using System.IO;
-using System;
-using ASI.Basecode.Services.Interfaces;
-using ASI.Basecode.Data.Models;
-
-namespace ASI.Basecode.WebApp.Controllers
-{
-    public class UserController : Controller
-    {
-        private readonly IUserService _userService;
-        public UserController(IUserService userService)
-        {
-            _userService = userService;
-        }
-
-        // GET: UsersController
-        public ActionResult UserList()
-        {
-            var users = _userService.GetAllUsers(); // Fetch users from the service
-            return View(users);
-        }
-
-        // GET: UsersController/UserAdd
-        public ActionResult UserAdd()
-        {
-            return View();
-        }
-
-        // GET: UsersController/Edit/UserId
-        public ActionResult Edit(string id)
-        {
-            var user = _userService.GetUserById(id);
-            if (user == null)
-            {
-                return NotFound();
-            }
-            return View(user);
-        }
-
-        // POST: UsersController/Edit/userId
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public ActionResult Edit(User user)
-        {
-            try
-            {
-                if (ModelState.IsValid)
-                {
-                    _userService.UpdateUser(user);
-                    return RedirectToAction(nameof(UserList));
-                }
-                return View(user);
-            }
-            catch
-            {
-                return View(user);
-            }
-        }
-
-        // POST: UsersController/ResetPassword/userId
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public ActionResult ResetPassword(string id)
-        {
-            try
-            {
-                _userService.ResetPassword(id, "Temp_123");
-                return RedirectToAction(nameof(UserList));
-            }
-            catch
-            {
-                return View();
-            }
-        }
-
-        // GET: UsersController/Delete/userId
-        public ActionResult Delete(string id)
-        {
-            try
-            {
-                _userService.DeleteUser(id);
-                return RedirectToAction(nameof(UserList));
-            }
-            catch
-            {
-                return View();
-            }
-        }
-
-        // POST: UsersController/Delete/userId
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public ActionResult Delete(int id, IFormCollection collection)
-        {
-            try
-            {
-                return RedirectToAction(nameof(Index));
-            }
-            catch
-            {
-                return View();
-            }
-        }
-    }
-}
