+++ ASI.Basecode.WebApp/Views/Shared/_Layout.cshtml
@@ -14,12 +14,12 @@
     @RenderSection("Styles", required: false)
 </head>
 <body>
-    @Html.Partial("_SideBar")
-    <div class="container">
+    @* @Html.Partial("_SideBar") *@
+    @* <div class="container"> *@
         <main role="main">
             @RenderBody()
         </main>
-    </div>
+   @*  </div> *@
     @* <footer class="border-top footer text-muted">
         <div class="container">
             @Copyright
