--- ASI.Basecode.WebApp/Views/Shared/_SideBar.cshtml
@@ -1,84 +1,0 @@
-﻿@using Microsoft.AspNetCore.Http
-@inject IHttpContextAccessor HttpContextAccessor
-
-@if (User.Identity.IsAuthenticated)
-{
-    <div class="sidebar">
-        <div class="sidebar-header hover-pointer">
-            <img src="~/img/scsHelpDeskLogo.svg" style="width: 185px;" alt="logo"
-                 onclick="window.location.href = '@Url.Action("Index", "Home")';">
-        </div>
-        <ul class="nav-menu">
-            <li>
-                <a href="@Url.Action("SuperAdminDashboard", "SuperAdmin")">
-                    <i class="fa-solid fa-home"></i> DASHBOARD
-                </a>
-            </li>
-            <li>
-                <a href="@Url.Action("UserList", "User")">
-                    <i class="fa-solid fa-user"></i> USER LIST
-                </a>
-            </li>
-            <li>
-                <a href="@Url.Action("", "")">
-                    <i class="fa-solid fa-book"></i> KNOWLEDGE BASE
-                </a>
-            </li>
-            <li>
-                <a href="@Url.Action("", "")">
-                    <i class="fa-solid fa-ticket"></i> TICKETS
-                </a>
-            </li>
-            <li>
-                <a href="@Url.Action("Settings", "Account")">
-                    <i class="fa-solid fa-gear"></i> ACCOUNT SETTINGS
-                </a>
-            </li>
-            <li>
-                <a asp-action="SignOutUser" asp-controller="Account">
-                    <i class="fa-solid fa-right-from-bracket"></i> LOGOUT
-                </a>
-            </li>
-        </ul>
-    </div>
-}
-
-
-<style>
-    .sidebar {
-        width: 250px;
-        height: 100vh;
-        background-color: #fff;
-        position: fixed;
-        top: 0;
-        left: 0;
-        padding: 20px;
-        color: #fff;
-    }
-
-    .sidebar-header img {
-        display: block;
-        margin: 0 auto 30px;
-    }
-
-    .nav-menu {
-        list-style-type: none;
-        padding: 0;
-    }
-
-        .nav-menu li {
-            margin-bottom: 20px;
-        }
-
-        .nav-menu a {
-            color: #333;
-            text-decoration: none;
-            font-size: 18px;
-            display: flex;
-            align-items: center;
-        }
-
-        .nav-menu i {
-            margin-right: 10px;
-        }
-</style>
