+++ ASI.Basecode.WebApp/Views/SuperAdmin/UserAdd.cshtml
@@ -2,7 +2,35 @@
     ViewData["Title"] = "Add User - SCS Help Desk";
 }
 
-<div style="display: flex;">
+<div style="display: flex; min-height: 100vh; width: 100vw;">
+    <!-- Sidebar -->
+    <div style="width: 250px; background-color: #f8f9fa;">
+        <div style="padding: 15px;">
+            <img src="~/img/SCS-Logo.png" alt="SCS Logo" style="max-width: 50px; margin-bottom: 10px;">
+            <h5 style="margin-bottom: 20px;">SCS HELP DESK</h5>
+        </div>
+        <div>
+            <a href="@Url.Action("SuperAdminDashboard", "SuperAdmin")" style="display: block; padding: 10px 15px; text-decoration: none; color: black;">
+                <i class="bi bi-grid"></i> DASHBOARD
+            </a>
+            <a href="@Url.Action("UserList", "User")" style="display: block; padding: 10px 15px; text-decoration: none; color: white; background-color: #6c757d;">
+                <i class="bi bi-person"></i> USER LIST
+            </a>
+            <a href="@Url.Action("Index", "KnowledgeBase")" style="display: block; padding: 10px 15px; text-decoration: none; color: black;">
+                <i class="bi bi-book"></i> KNOWLEDGE BASE
+            </a>
+            <a href="@Url.Action("Index", "Tickets")" style="display: block; padding: 10px 15px; text-decoration: none; color: black;">
+                <i class="bi bi-ticket"></i> TICKETS
+            </a>
+            <a href="@Url.Action("Index", "AccountSettings")" style="display: block; padding: 10px 15px; text-decoration: none; color: black;">
+                <i class="bi bi-gear"></i> ACCOUNT SETTINGS
+            </a>
+            <a asp-action="SignOutUser" asp-controller="Account" style="display: block; padding: 10px 15px; text-decoration: none; color: black;">
+                <i class="bi bi-box-arrow-right"></i> LOGOUT
+            </a>
+        </div>
+    </div>
+
     <!-- Main Content -->
     <div style="flex: 1; background: linear-gradient(135deg, #ff6b00 0%, #ffa500 100%); padding: 20px;">
         <div style="background-color: white; border-radius: 15px; padding: 30px; max-width: 800px; margin: 0 auto;">
@@ -75,3 +103,7 @@
         </div>
     </div>
 </div>
+
+@section Styles {
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.7.2/font/bootstrap-icons.css">
+}
