+++ ASI.Basecode.WebApp/Views/SuperAdmin/UserList.cshtml
@@ -4,6 +4,35 @@
     ViewData["Title"] = "User List - SCS Help Desk";
 }
 
+<div style="display: flex; min-height: 100vh;">
+    <!-- Sidebar -->
+    <div style="width: 250px; background-color: #f8f9fa; ">
+        <div style="padding: 15px;">
+            <img src="~/img/SCS-Logo.png" alt="SCS Logo" style="max-width: 50px; margin-bottom: 10px;">
+            <h5 style="margin-bottom: 20px;">SCS HELP DESK</h5>
+        </div>
+        <div>
+            <a href="@Url.Action("SuperAdminDashboard", "SuperAdmin")" style="display: block; padding: 10px 15px; text-decoration: none; color: black;">
+                <i class="bi bi-grid"></i> DASHBOARD
+            </a>
+            <a href="@Url.Action("UserList", "SuperAdmin")" style="display: block; padding: 10px 15px; text-decoration: none; color: white; background-color: #6c757d;">
+                <i class="bi bi-person"></i> USER LIST
+            </a>
+            <a href="@Url.Action("Index", "KnowledgeBase")" style="display: block; padding: 10px 15px; text-decoration: none; color: black;">
+                <i class="bi bi-book"></i> KNOWLEDGE BASE
+            </a>
+            <a href="@Url.Action("Index", "Tickets")" style="display: block; padding: 10px 15px; text-decoration: none; color: black;">
+                <i class="bi bi-ticket"></i> TICKETS
+            </a>
+            <a href="@Url.Action("Index", "AccountSettings")" style="display: block; padding: 10px 15px; text-decoration: none; color: black;">
+                <i class="bi bi-gear"></i> ACCOUNT SETTINGS
+            </a>
+            <a asp-action="SignOutUser" asp-controller="Account" style="display: block; padding: 10px 15px; text-decoration: none; color: black;">
+                <i class="bi bi-box-arrow-right"></i> LOGOUT
+            </a>
+        </div>
+    </div>
+
 <div style="flex: 1; padding: 20px;">
     <!-- Main Content -->
     <div style="background-color: white; border-radius: 15px; padding: 30px;">
@@ -27,7 +56,7 @@
                         </select>
                     </div>
                     <div class="col-md-4 text-end">
-                        <a href="@Url.Action("UserAdd", "User")" class="btn btn-warning">
+                        <a href="@Url.Action("UserAdd", "SuperAdmin")" class="btn btn-warning">
                             <i class="bi bi-plus"></i> Add User
                         </a>
                     </div>
@@ -53,7 +82,7 @@
                                     <td>@user.UserId</td>
                                     <td>@user.Email</td>
                                     <td>
-                                        <a href="@Url.Action("Edit", "User", new { id = user.UserId })" class="btn btn-primary">Edit</a>
+                                        <a href="@Url.Action("Edit", "SuperAdmin", new { id = user.UserId })" class="btn btn-primary">Edit</a>
                                         <a href="#" class="btn btn-danger" onclick="confirmDelete('@user.UserId')">Delete</a>
                                     </td>
                                 </tr>
@@ -70,8 +99,11 @@
     <script>
         function confirmDelete(userId) {
             if (confirm("Are you sure you want to delete this user?")) {
-                window.location.href = '@Url.Action("Delete", "User")?id=' + userId;
+                window.location.href = '@Url.Action("Delete", "SuperAdmin")?id=' + userId;
             }
         }
     </script>
 }
+@section Styles {
+        <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.7.2/font/bootstrap-icons.css">
+    }
